feature/add-compare-methods-to-math-helper
- Reworking a bit.
